feat: CRUD comments terminado
Se crea el CRUD completo de comentarios, ademas se agrego validaciones para que solo el autor del comentario y publicación pueda eliminar y editar

+++ app-viajes-backend/src/controllers/post.controller.js
@@ -87,6 +87,13 @@ export const updatePost = async (req, res) => {
     }
     const { postId } = req.params;
     const publicacion = req.body;
+    const publicacionEncontrada = await PostModel.findById(postId);
+    // Verifica si el publicacion existe
+    if (!publicacionEncontrada) return res.status(404).json({ message: "Comentario no encontrado." });
+    // Verifica si el usuario actual es el autor de la publicacion
+    if (!publicacionEncontrada.author.equals(req.userId)) {
+      return res.status(403).json({ message: "No tienes permisos para editar este comentario." });
+    }
     const publicacionActualizada = await PostModel.findByIdAndUpdate(postId, publicacion, { new: true });
     res.status(202).json(publicacionActualizada);
   } catch (error) {
@@ -99,6 +106,13 @@ export const updatePost = async (req, res) => {
 export const deletePost = async (req, res) => {
   try {
     const { postId } = req.params;
+    const publicacionEncontrada = await PostModel.findById(postId);
+    // Verifica si el publicacion existe
+    if (!publicacionEncontrada) return res.status(404).json({ message: "Publicación no encontrado." });
+    // Verifica si el usuario actual es el autor de la publicacion
+    if (!publicacionEncontrada.author.equals(req.userId)) {
+      return res.status(403).json({ message: "No tienes permisos para editar este comentario." });
+    }
     await PostModel.findByIdAndDelete(postId);
     res.status(200).json({ message: "Publicación eliminada" });
   } catch (error) {
